+++ vector/src/main/java/im/vector/app/features/home/room/detail/RoomDetailFragment.kt
@@ -655,6 +655,14 @@ class RoomDetailFragment @Inject constructor(
                 roomDetailViewModel.handle(RoomDetailAction.ClearSendQueue)
                 true
             }
+            R.id.invite              -> {
+                navigator.openInviteUsersToRoom(requireActivity(), roomDetailArgs.roomId)
+                true
+            }
+            R.id.timeline_setting    -> {
+                navigator.openRoomProfile(requireActivity(), roomDetailArgs.roomId)
+                true
+            }
             R.id.resend_all          -> {
                 roomDetailViewModel.handle(RoomDetailAction.ResendAll)
                 true