+++ vector/src/main/java/im/vector/app/features/home/room/detail/RoomDetailViewState.kt
@@ -72,6 +72,7 @@ data class RoomDetailViewState(
         val canShowJumpToReadMarker: Boolean = true,
         val changeMembershipState: ChangeMembershipState = ChangeMembershipState.Unknown,
         val canSendMessage: Boolean = true,
+        val canInvite: Boolean = true,
         val isAllowedToManageWidgets: Boolean = false,
         val isAllowedToStartWebRTCCall: Boolean = true
 ) : MvRxState {